+++ api/rest/dao/ResultDao.class.php
@@ -3,57 +3,59 @@
 require_once __DIR__.'/BaseDao.class.php';
 
 class ResultDao extends BaseDao {
+    private $userID;
 
-    public function __construct() {
-        parent::__construct("result");
+    public function __construct($userID = NULL) {
+        parent::__construct("result", $userID);
+        $this->userID = $userID;
     }
 
     public function getAllresults() {
         return $this->query("SELECT r.resultID, r.clubMemberID, m.firstName, m.lastName, r.opponentFirstName, r.opponentLastName, r.resultStatus, r.appUserID, r.tournamentID
-        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID;", []);
+        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID having r.appUserID = :appUserID;", ["appUserID" => $this->userID]);
     }
 
-    public function getresults($offset, $limit, $order) {
+    public function getResults($offset, $limit, $order) {
         list($order_column, $order_direction) = self::parseOrder($order);
 
         return $this->query("SELECT r.resultID, r.clubMemberID, m.firstName, m.lastName, r.opponentFirstName, r.opponentLastName, r.resultStatus, r.appUserID, r.tournamentID
-        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID ORDER BY {$order_column} {$order_direction}
-                           LIMIT {$limit} OFFSET {$offset} ;", []);
+        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID having appUserID = :appUserID ORDER BY {$order_column} {$order_direction}
+                           LIMIT {$limit} OFFSET {$offset} ;", ["appUserID" => $this->userID]);
     }
 
-    public function getresultByID($id) {
+    public function getResultByID($id) {
         return $this->query("SELECT r.resultID, r.clubMemberID, m.firstName, m.lastName, r.opponentFirstName, r.opponentLastName, r.resultStatus, r.appUserID, r.tournamentID
-        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID having r.resultID = :resultID;", ["resultID" => $id]);
+        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID having r.resultID = :resultID and r.appUserID = :appUserID;", ["resultID" => $id, "appUserID" => $this->userID]);
     }
 
-    public function addresult($result) {
+    public function addResult($result) { 
         return $this->add($result);
     }
 
-    public function updateresult($id, $result) {
+    public function updateResult($id, $result) {
         return $this->update($id, $result, "resultID");
     }
 
-    public function deleteresult($id) {
+    public function deleteResult($id) {
         return $this->delete($id, "resultID");
     }
 
     public function getResultsByTournamentID($tournamentID) {
         return $this->query("SELECT r.resultID, r.clubMemberID, m.firstName, m.lastName, r.opponentFirstName, r.opponentLastName, r.resultStatus, r.appUserID, r.tournamentID
-        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID having r.tournamentID = :tournamentID;", ["tournamentID" => $tournamentID]);
+        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID having r.tournamentID = :tournamentID and r.appUserID = :appUserID;", ["tournamentID" => $tournamentID, "appUserID" => $this->userID]);
     }
 
     public function getResultsByClubMemberID($clubMemberID) {
         return $this->query("SELECT r.resultID, r.clubMemberID, m.firstName, m.lastName, r.opponentFirstName, r.opponentLastName, r.resultStatus, r.appUserID, r.tournamentID
-        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID having r.clubMemberID = :clubMemberID;", ["clubMemberID" => $clubMemberID]);    
+        from `result` r join clubMember m on r.clubMemberID = m.clubMemberID having r.clubMemberID = :clubMemberID and r.appUserID = :appUserID;", ["clubMemberID" => $clubMemberID, "appUserID" => $this->userID]);    
     }
 
     public function deleteResultsForTournament($tournamentID) {
-        return $this->query("DELETE FROM result WHERE tournamentID = :tournamentID", ["tournamentID" => $tournamentID]);
+        return $this->query("DELETE FROM result WHERE tournamentID = :tournamentID and appUserID = :appUserID", ["tournamentID" => $tournamentID, "appUserID" => $this->userID]);
     }
 
     public function deleteResultsForMember($clubMemberID) {
-        return $this->query("DELETE FROM result WHERE clubMemberID = :clubMemberID", ["clubMemberID" => $clubMemberID]);
+        return $this->query("DELETE FROM result WHERE clubMemberID = :clubMemberID and appUserID = :appUserID", ["clubMemberID" => $clubMemberID, "appUserID" => $this->userID]);
     }
 }
